+++ src/main/java/online/kingdomkeys/kingdomkeys/client/render/entity/ElementalMusicalHeartlessRenderer.java
@@ -5,7 +5,9 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.MobRenderer;
 import net.minecraft.resources.ResourceLocation;
+import online.kingdomkeys.kingdomkeys.KingdomKeys;
 import online.kingdomkeys.kingdomkeys.client.model.entity.ElementalMusicalHeartlessModel;
+import online.kingdomkeys.kingdomkeys.client.render.HeartlessEyesLayerRenderer;
 import online.kingdomkeys.kingdomkeys.entity.mob.BaseElementalMusicalHeartlessEntity;
 import online.kingdomkeys.kingdomkeys.entity.mob.EmeraldBluesEntity;
 
@@ -15,6 +17,7 @@ public class ElementalMusicalHeartlessRenderer extends MobRenderer<BaseElemental
 
     public ElementalMusicalHeartlessRenderer(EntityRendererProvider.Context context) {
         super(context, new ElementalMusicalHeartlessModel<>(context.bakeLayer(ElementalMusicalHeartlessModel.LAYER_LOCATION)), 0.35F);
+        this.addLayer(new HeartlessEyesLayerRenderer<>(this, ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, "textures/entity/mob/musical_heartless_eyes.png")));
     }
 
     @Override