+++ src/main/java/online/kingdomkeys/kingdomkeys/integration/epicfight/PatchedDriveLayerRenderer.java
@@ -10,6 +10,7 @@ import net.minecraft.client.resources.PlayerSkin;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import online.kingdomkeys.kingdomkeys.KingdomKeys;
 import online.kingdomkeys.kingdomkeys.config.ModConfigs;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.driveform.DriveForm;
@@ -23,6 +24,7 @@ import yesman.epicfight.gameasset.Armatures;
 import yesman.epicfight.world.capabilities.entitypatch.LivingEntityPatch;
 
 public class PatchedDriveLayerRenderer<E extends LivingEntity, T extends LivingEntityPatch<E>, M extends EntityModel<E>> extends PatchedLayer<E, T, M, RenderLayer<E, M>> {
+    private static final ResourceLocation ANTI_EYES = ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, "textures/models/armor/anti_eyes.png");
 
     public PatchedDriveLayerRenderer() {  }
 
@@ -32,9 +34,12 @@ public class PatchedDriveLayerRenderer<E extends LivingEntity, T extends LivingE
             String drive = PlayerData.get((Player) entity).getActiveDriveForm();
             DriveForm form = ModDriveForms.registry.get(ResourceLocation.parse(drive));
             if (form.getTextureLocation((Player) entity) != null) {
-                //VertexConsumer vertexConsumer = EpicFightRenderTypes.getArmorFoilBufferTriangles(multiBufferSource, RenderType.armorCutoutNoCull(form.getTextureLocation((Player) e)), true, false);
                 HumanoidMesh model = getModel(entity);
-                model.draw(poseStack, multiBufferSource, EpicFightRenderTypes.armorCutoutNoCull(form.getTextureLocation((Player) entity)), i, 1, 1, 1, 1, OverlayTexture.NO_OVERLAY, Armatures.BIPED.get(), openMatrix4fs);
+                model.draw(poseStack, multiBufferSource, EpicFightRenderTypes.entityCutoutNoCull(form.getTextureLocation((Player) entity)), i, 1, 1, 1, 1, OverlayTexture.NO_OVERLAY, Armatures.BIPED.get(), openMatrix4fs);
+
+                if(form.equals(ModDriveForms.ANTI.get())) {
+                    model.draw(poseStack, multiBufferSource, EpicFightRenderTypes.eyes(ANTI_EYES), 15728640, 1, 1, 1, 1, OverlayTexture.NO_OVERLAY, Armatures.BIPED.get(), openMatrix4fs);
+                }
             }
 
         }