+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+**Added:**
+* Desertor heartless.
+
+**Changed:**
+* Heartless and Antiform eyes will now glow in the darkness.
+* Font and color used in the XP obtained from an enemy.
+
 **Fixed:**
-* Rendering issues caused by Drive bar fixes icons not rendering with the Tool Belt mod and potentially other issues
+*